# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tên API của bạn",
    "version": "1.0.0",
    "description": "Tài liệu mô tả API sử dụng Swagger UI",
    "contact": {
      "name": "Tên bạn hoặc nhóm phát triển",
      "email": "email@example.com"
    }
  },
  "servers": [
    {
      "url": "fe-api.training.ssit.company",
      "description": "Server testing"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "StockMovement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID của stock movement",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "description": "ID sản phẩm",
            "example": "abc12345-e89b-12d3-a456-426614174003"
          },
          "type": {
            "type": "string",
            "enum": [
              "IN",
              "OUT",
              "ADJUSTMENT"
            ],
            "description": "Loại movement",
            "example": "IN"
          },
          "quantity": {
            "type": "integer",
            "description": "Số lượng thay đổi",
            "example": 50
          },
          "beforeQuantity": {
            "type": "integer",
            "description": "Số lượng trước khi thay đổi",
            "example": 100
          },
          "afterQuantity": {
            "type": "integer",
            "description": "Số lượng sau khi thay đổi",
            "example": 150
          },
          "reason": {
            "type": "string",
            "description": "Lý do thay đổi",
            "example": "Nhập hàng từ nhà cung cấp"
          },
          "notes": {
            "type": "string",
            "description": "Ghi chú",
            "example": "Hàng chất lượng tốt"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "ID người thực hiện",
            "example": "456e7890-e89b-12d3-a456-426614174001"
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "description": "ID đơn hàng (nếu có)",
            "example": "789e1234-e89b-12d3-a456-426614174002"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Ngày tạo",
            "example": "2025-01-20T14:30:00.000Z"
          }
        }
      },
      "StockAdjustmentRequest": {
        "type": "object",
        "required": [
          "productId",
          "newQuantity",
          "reason"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid",
            "description": "ID sản phẩm",
            "example": "abc12345-e89b-12d3-a456-426614174003"
          },
          "newQuantity": {
            "type": "integer",
            "minimum": 0,
            "description": "Số lượng mới",
            "example": 120
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Lý do điều chỉnh",
            "example": "Kiểm kê định kỳ"
          },
          "notes": {
            "type": "string",
            "maxLength": 500,
            "description": "Ghi chú",
            "example": "Phát hiện hàng bị hỏng"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/stock/movements": {
      "get": {
        "summary": "Lấy danh sách tất cả movement kho",
        "description": "Lấy danh sách movement kho với phân trang và tìm kiếm",
        "tags": [
          "Stock"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Số lượng items mỗi trang"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "IN",
                "OUT",
                "ADJUSTMENT"
              ]
            },
            "description": "Lọc theo loại movement"
          },
          {
            "in": "query",
            "name": "productId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Lọc theo sản phẩm"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Từ ngày (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Đến ngày (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách movement kho",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "movements": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/StockMovement"
                              }
                            },
                            "pagination": {
                              "$ref": "#/components/schemas/PaginationMeta"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/adjustment": {
      "post": {
        "summary": "Tạo stock adjustment",
        "description": "Tạo điều chỉnh tồn kho cho sản phẩm",
        "tags": [
          "Stock"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockAdjustmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo adjustment thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StockMovement"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy sản phẩm",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/movements/{id}": {
      "get": {
        "summary": "Lấy thông tin movement kho theo ID",
        "description": "Lấy thông tin chi tiết của một movement kho dựa trên ID",
        "tags": [
          "Stock"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID của movement"
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin movement kho",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StockMovement"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "ID không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy movement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/movements/today": {
      "get": {
        "summary": "Lấy movement kho hôm nay",
        "description": "Lấy danh sách tất cả movement kho trong ngày hôm nay",
        "tags": [
          "Stock"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách movement hôm nay",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StockMovement"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/stats": {
      "get": {
        "summary": "Lấy thống kê kho",
        "description": "Lấy thống kê tổng quan về tình trạng kho hàng",
        "tags": [
          "Stock"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thống kê kho hàng",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StockStats"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/daily-report": {
      "get": {
        "summary": "Lấy báo cáo hàng ngày",
        "description": "Lấy báo cáo hoạt động kho hàng trong ngày",
        "tags": [
          "Stock"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Ngày báo cáo (YYYY-MM-DD), mặc định là hôm nay"
          }
        ],
        "responses": {
          "200": {
            "description": "Báo cáo hoạt động kho hàng",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DailyStockReport"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Chưa đăng nhập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền truy cập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}